A gt theme for Blanco Center tables

gt_theme_bc() adds the Blanco Center theme to any table created using gt(). Customize table elements like font family, or text color and sizes.

Usage

gt_theme_bc(
  gt,
  base_family = "Quicksand",
  header_family = "Bitter",
  text_color = "#575E66",
  title_color = "#a00b0b",
  base_size = 12,
  ...
)

Arguments

gt

An existing gt table object of class gt_tbl

base_family

Font used for most text.

header_family

Font used for titles: table title, column headers, spanner column headers, row group labels.

text_color

Color used for everything else, except table title.

title_color

Color used for table title.

base_size

Base text size in pts (converted to px within function).

...

Additional arguments passed to gt::tab_options().

Value

A gt theme.

Examples

if (FALSE) { # \dontrun{
gt_theme_bc(header_family = "Bitter", base_family = "Quicksand")
} # }