Louisiana's 64 parishes, FIPS codes, and region designations according to different state agencies and organizations, parish geometry, and parish centers.

Louisiana's 64 parishes, FIPS codes, and region designations according to different state agencies and organizations, parish geometry, and parish centers.

Usage

louisiana

Format

A data frame/sf object with 64 rows and 17 columns:

GEOID, FIPS

Parish FIPS code in character & numeric formats

NAME

Full parish name including "Parish, Louisiana"

variable

Variable name for population density, consists of Census table and row number

estimate

Estimate

moe

Margin of error

year

Year for which ACS-5 Year estimates were retrieved

parish, parish_full

Parish name with & without "Parish"

region_DE, region_DCFS, region_laborforce, region_LDH, region_OJJ

Region numeric ID

region_name_DE, region_name_DCFS, region_name_laborforce, region_BLS, region_name_LDH, region_name_OJJ

Region name ID

one_acadiana

Designates parishes within Acadiana region

metro_BLS

"Metropolitan area" or "Nonmetropolitan area" as designated by BLS

geometry

MULTIPOLYGON, parish outline

pop_density

Rural/Urban, threshold = 100000

center

POINT, parish center point

Source

U.S. Census Bureau, 'tidycensus' package, see sources for 'parishes'