A ggplot2 theme for Blanco Center visualizations

theme_bc_legacy() adds the Blanco Center theme to any plot created using ggplot(). Customize plot elements like font family and color and axis colors and line sizes. Grid lines for x and y axes can be added, or all axes can be removed, e.g. for maps. Note: This is the legacy version of the Blanco Center theme function (developed in 2023). For the updated version, use theme_bc(), developed Fall 2025.

Usage

theme_bc_legacy(
  title_font = "Bitter",
  base_font = "Quicksand",
  dark_text = "#1A242F",
  light_text = "#575E66",
  void = FALSE,
  grid_x = FALSE,
  grid_y = FALSE,
  base_size = 12,
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

Arguments

title_font

Font used for titles: plot title, axis titles, legend title.

base_font

Font used for all other text.

dark_text

Color used for axis and legend titles.

light_text

Color used for everything else, except plot title.

void

TRUE or FALSE: removes all axis elements. Suggested for maps.

grid_x

TRUE or FALSE: adds x gridlines.

grid_y

TRUE or FALSE: adds y gridlines.

base_size

Base text size in pts.

base_line_size

Base line size in pts.

base_rect_size

Base rectangle size in pts.

Value

A ggplot2 theme.

Examples

if (FALSE) { # \dontrun{
theme_bc_legacy(title_font = "Bitter", base_font = "Quicksand", void = TRUE)
} # }